refactor(quality): split modules to cut file-level complexity below Codacy goal
Codacy's complexFilesPercentage was ~16% (23/145 files over its default
file-complexity threshold of 40) against the <=10% goal. Split eleven of the
most-complex scripts/quality modules along cohesive seams (v1/v2 normalisers,
report writer, GitHub health probes, prompt section renderers, migration
rules, rollup logic, alternate-view renderers, Codacy pending-message
helpers), re-exporting the moved public names so import surfaces are
unchanged. No metric gaming, suppressions, or threshold changes: each split
genuinely reduces aggregate per-file cyclomatic complexity. Result: 14/157
files (8.9%) over threshold; max per-function CCN stays 8.

Wire the new traced helper (migrate_profiles_to_v2_rules) into
tool.coverage.run.source and add the test-excluded helpers to
sonar.coverage.exclusions so the coverage-source/exclusion contract holds.
100% line+branch coverage maintained (1509 tests pass).
