+++ scripts/quality/migrate_profiles_to_v2.py
@@ -28,139 +28,18 @@ from __future__ import absolute_import  # noqa: UP010 — required by codacy-com
 import argparse
 import sys
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Mapping
+from typing import Iterable, List
 
 if str(Path(__file__).resolve().parents[2]) not in sys.path:  # pragma: no cover
     sys.path.insert(0, str(Path(__file__).resolve().parents[2]))
 
 import yaml  # type: ignore[import-untyped]
 
-from scripts.quality.profile_normalization import (
-    normalize_mode,
-    normalize_scanners,
+from scripts.quality.migrate_profiles_to_v2_rules import (  # noqa: F401  pylint: disable=unused-import
+    migrate_profile,
 )
 
 PROFILES_DIR_DEFAULT = Path(__file__).resolve().parents[2] / "profiles" / "repos"
-
-
-def migrate_profile(raw_profile: Mapping[str, Any]) -> Dict[str, Any]:
-    """Return a v2-migrated copy of ``raw_profile`` without mutating input.
-
-    Keys already present at their v2 values are left untouched. Additions
-    are inserted in a canonical order so the on-disk diff is small and
-    reviewable.
-    """
-    migrated: Dict[str, Any] = dict(raw_profile)
-    _ensure_version(migrated)
-    _ensure_mode(migrated)
-    _ensure_scanners(migrated)
-    _ensure_overrides(migrated)
-    _ensure_coverage_flags(migrated)
-    return _reorder_for_review(migrated)
-
-
-def _ensure_version(migrated: Dict[str, Any]) -> None:
-    """Set ``version: 2`` unless the file already declares v2 or later."""
-    if "version" not in migrated:
-        migrated["version"] = 2
-        return
-    # Clamp unknown / v1 declarations to 2 — once this migrator touches a
-    # file, the file is v2 by definition.
-    try:
-        declared = int(migrated["version"])
-    except (TypeError, ValueError):
-        declared = 1
-    if declared < 2:
-        migrated["version"] = 2
-
-
-def _ensure_mode(migrated: Dict[str, Any]) -> None:
-    """Synthesise ``mode`` from legacy ``issue_policy`` when absent."""
-    if "mode" not in migrated:
-        migrated["mode"] = _mode_block(migrated.get("issue_policy"))
-
-
-def _ensure_scanners(migrated: Dict[str, Any]) -> None:
-    """Synthesise ``scanners`` from legacy ``enabled_scanners`` when absent."""
-    if "scanners" not in migrated:
-        migrated["scanners"] = _scanners_block(migrated.get("enabled_scanners"))
-
-
-def _ensure_overrides(migrated: Dict[str, Any]) -> None:
-    """Ensure ``overrides`` exists as (at minimum) an empty list."""
-    if "overrides" not in migrated:
-        migrated["overrides"] = []
-
-
-def _ensure_coverage_flags(migrated: Dict[str, Any]) -> None:
-    """Rewrite ``coverage.inputs`` so each item carries a ``flag``."""
-    coverage = migrated.get("coverage")
-    if not isinstance(coverage, Mapping):
-        return
-    inputs = coverage.get("inputs")
-    if not isinstance(inputs, list):
-        return
-    migrated_coverage = dict(coverage)
-    migrated_coverage["inputs"] = _migrate_coverage_inputs(inputs)
-    migrated["coverage"] = migrated_coverage
-
-
-def _mode_block(legacy_issue_policy: Any) -> Dict[str, Any]:
-    """Produce a tidy ``mode`` block for on-disk YAML.
-
-    Uses the shared normaliser (same logic as the load-time path) and
-    drops the ratchet sub-block when the phase isn't ratchet, so the
-    YAML stays compact and readable.
-    """
-    result = normalize_mode(None, legacy_issue_policy=legacy_issue_policy)
-    if result["phase"] != "ratchet":
-        # Drop the placeholder ratchet + shadow_until fields for non-ratchet
-        # repos so the YAML diff stays minimal. They can be re-added later
-        # if the repo ever needs them.
-        return {"phase": result["phase"]}
-    return {
-        "phase": result["phase"],
-        "ratchet": result["ratchet"],
-    }
-
-
-def _scanners_block(legacy_enabled_scanners: Any) -> Dict[str, Dict[str, str]]:
-    """Emit an explicit ``scanners`` map derived from legacy enabled flags."""
-    return normalize_scanners(None, legacy_enabled_scanners=legacy_enabled_scanners)
-
-
-def _migrate_coverage_inputs(inputs: List[Any]) -> List[Dict[str, Any]]:
-    """Ensure every coverage input carries a ``flag`` (from ``name``)."""
-    migrated: List[Dict[str, Any]] = []
-    for item in inputs:
-        if not isinstance(item, Mapping):
-            migrated.append(item)  # type: ignore[arg-type]
-            continue
-        entry = dict(item)
-        if "flag" not in entry or not str(entry.get("flag", "")).strip():
-            entry["flag"] = str(entry.get("name", "")).strip()
-        migrated.append(entry)
-    return migrated
-
-
-def _reorder_for_review(profile: Dict[str, Any]) -> Dict[str, Any]:
-    """Emit v2 fields near the top so YAML diffs land where a reviewer looks."""
-    preferred_order = [
-        "slug",
-        "version",
-        "stack",
-        "mode",
-        "scanners",
-        "overrides",
-    ]
-    ordered: Dict[str, Any] = {}
-    for key in preferred_order:
-        if key in profile:
-            ordered[key] = profile[key]
-    for key, value in profile.items():
-        if key not in ordered:
-            ordered[key] = value
-    return ordered
 
 
 def _iter_profile_paths(profiles_dir: Path) -> Iterable[Path]:
